pyproject.toml
[project]
name = "bound-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.2",
    "asyncpg>=0.31.0",
    "boto3>=1.42.37",
    "fastapi[standard]>=0.128.0",
    "gunicorn>=23.0.0",
    "greenlet>=3.3.1",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    # TTS dependencies
    "pydub>=0.25.1",
    "pysrt>=1.1.2",
    "librosa>=0.10.0",
    "soundfile>=0.12.1",
    "openai-whisper>=20231117",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.4.2",
    "aiosqlite>=0.21.0",
    "ruff>=0.14.14",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
